Casting and RNG Hot Fix (#127)
* Add some latex for MathJax testing.

* Improve forward and backward Javadoc.

* Texify Everything (#124)

Adds the ability to specify LaTeX in Javadoc.

Adds a GitHub action to automatically parse and inject latex into Javadocs published on the GitHub pages website.

* Convert all angles in Rotation.java to radians to be consistent.

* Improved Matrix Market support.

- Added support for dense/coo symmetric/Hermitian matrices (reading and writing).
- Added support for coo pattern matrices (reading and writing.

* Improve random generation and package naming

* Improve random sparse tensor generation

* Add configurable dispatcher for complex matrices

* Fix latex comments

* Fix down-casting and rng issues

+++ src/main/java/org/flag4j/linalg/ops/dispatch/ReDeMatMultDispatcher.java
@@ -156,7 +156,7 @@ public final class ReDeMatMultDispatcher extends BiTensorOpDispatcher<Matrix, Ma
      */
     public static Matrix dispatch(Matrix a, Matrix b) {
         // Use a long to protect against possible overflow.
-        long totalEntries = a.dataLength() + b.dataLength();
+        long totalEntries = (long) a.dataLength() + b.dataLength();
 
         // Return as fast as possible for "small-enough" matrices.
         if(totalEntries < SML_THRESH) {
@@ -188,7 +188,7 @@ public final class ReDeMatMultDispatcher extends BiTensorOpDispatcher<Matrix, Ma
 
         if(k == 1) {
             // Then we have a matrix-vector product.
-            double aspectRatio = Math.max(m, n) / Math.min(m, n);
+            double aspectRatio = (double) Math.max(m, n) / Math.min(m, n);
 
             if(aspectRatio <= 4*ASPECT_THRESH) {
                 // The matrix is approximately square.
